SQUASH: apis: move & rename tenancy ClusterWorkspaceShard to core Shard – fix gvr in cache replicator
